requirements/production.txt
django
requests
django-pipeline
python-social-auth
hashids
djangorestframework
